fix: improve stop-level reliability chart mobile responsiveness and DTO type safety
- Add mobile optimizations to chart controller for stop-level reliability chart
  - Detect screen size (mobile < 768px, small mobile < 480px)
  - Adjust grid layout to give 50-60% space for stop names on mobile
  - Reduce font sizes for better mobile readability
  - Truncate long stop names with ellipsis
  - Auto-adjust on window resize and orientation changes

- Increase chart height on mobile (500px) for better vertical space usage
- Keep desktop height at 400px using responsive classes

- Fix type mismatch in BunchingByWeatherStatsDto
  - Change snowHours, rainHours, clearHours from int to float
  - Matches actual type from BunchingRateDto.exposureHours

- Replace generic arrays with typed DTOs in WeatherImpactDto
  - Use WinterOperationsStatsDto instead of array
  - Use TemperatureThresholdStatsDto instead of array
  - Use WeatherImpactMatrixStatsDto instead of array
  - Use BunchingByWeatherStatsDto instead of array
  - Provides full type safety for weather analysis insights

+++ src/Dto/Insight/BunchingByWeatherStatsDto.php
@@ -14,11 +14,11 @@ final readonly class BunchingByWeatherStatsDto implements \JsonSerializable
     public function __construct(
         public bool $hasData,
         public float $snowRate,
-        public int $snowHours,
+        public float $snowHours,
         public float $rainRate,
-        public int $rainHours,
+        public float $rainHours,
         public float $clearRate,
-        public int $clearHours,
+        public float $clearHours,
         public float $multiplier,
     ) {
     }
